+++ sdk/metorial/src/index.ts
@@ -1,1 +1,2 @@
 export * from '@metorial/sdk';
+export { default } from '@metorial/sdk';