+++ sdk/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @metorial/core
 
+## 2.1.0
+
+### Minor Changes
+
+- Change withProviderSession flow to direct session use
+
 ## 2.0.1
 
 ### Patch Changes