+++ sdk/metorial/CHANGELOG.md
@@ -1,5 +1,17 @@
 # metorial
 
+## 2.1.1
+
+### Patch Changes
+
+- Fix withProviderSession to work with adapter resolve method
+
+## 2.1.0
+
+### Minor Changes
+
+- Change withProviderSession flow to direct session use
+
 ## 2.0.1
 
 ### Patch Changes